refactor(RandomIntGenerator): util 패키지에서 model 패키지로 이동
