+++ src/main/java/racingcar/model/intgenerator/IntGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.model.intgenerator;
 
 public interface IntGenerator {
     int pickNumber();
