+++ src/main/java/racingcar/model/intgenerator/RandomIntGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.model.intgenerator;
 
 import camp.nextstep.edu.missionutils.Randoms;
