+++ src/test/java/racingcar/model/CarsTest.java
@@ -7,7 +7,7 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racingcar.util.CustomIntGenerator;
+import racingcar.model.intgenerator.CustomIntGenerator;
 
 public class CarsTest {
     @Test
